+++ module.h
@@ -8,7 +8,7 @@ public:
     Module();
     void enableParameter(int index);
     void disableParameter(int index);
-    void setParameter(int index, float value);
+    void setParameter(int index, double value);
     bool getParameterState(int index);
     double getParameterValue(int index);
     bool* getParameterStatesTable();
